ADD: MTL support

+++ finetune/comparison.py
@@ -29,7 +29,7 @@ class ComparisonPipeline(ClassificationPipeline):
 
     def feed_shape_type_def(self):
         TS = tf.TensorShape
-        return ({"tokens": tf.int32, "mask": tf.int32}, tf.int32), (
+        return ({"tokens": tf.int32, "mask": tf.float32}, tf.float32), (
             {"tokens": TS([2, self.config.max_length, 2]), "mask": TS([2, self.config.max_length])},
             TS([self.target_dim]))
